DDP-7047 | reorganize files into packages
