--- src/main/java/org/broadinstitute/dsm/model/elastic/export/Generator.java
@@ -1,9 +1,0 @@
-package org.broadinstitute.dsm.model.elastic.export;
-
-import java.util.Map;
-
-import org.broadinstitute.dsm.model.NameValue;
-
-public interface Generator {
-    Map<String, Object> generate();
-}
