+++ src/main/java/org/broadinstitute/dsm/model/elastic/export/generate/BaseGenerator.java
@@ -1,10 +1,11 @@
-package org.broadinstitute.dsm.model.elastic.export;
+package org.broadinstitute.dsm.model.elastic.export.generate;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
 import org.broadinstitute.dsm.db.structure.DBElement;
 import org.broadinstitute.dsm.model.NameValue;
 import org.broadinstitute.dsm.model.elastic.Util;
+import org.broadinstitute.dsm.model.elastic.export.parse.Parser;
 
 import java.util.Map;
 import java.util.Objects;
