+++ src/main/java/org/broadinstitute/dsm/model/elastic/export/generate/Generator.java
@@ -1,0 +1,7 @@
+package org.broadinstitute.dsm.model.elastic.export.generate;
+
+import java.util.Map;
+
+public interface Generator {
+    Map<String, Object> generate();
+}
