+++ src/main/java/org/broadinstitute/dsm/model/elastic/export/parse/TypeParser.java
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsm.model.elastic.export;
+package org.broadinstitute.dsm.model.elastic.export.parse;
 
 public class TypeParser extends BaseParser {
