+++ src/main/java/org/broadinstitute/dsm/model/patch/PatchFactory.java
@@ -1,10 +1,6 @@
 package org.broadinstitute.dsm.model.patch;
 
 import org.apache.commons.lang3.StringUtils;
-import org.broadinstitute.dsm.model.elastic.export.ElasticExportAdapter;
-import org.broadinstitute.dsm.model.elastic.export.MappingGenerator;
-import org.broadinstitute.dsm.model.elastic.export.SourceGenerator;
-import org.broadinstitute.dsm.model.elastic.export.TypeParser;
 import org.broadinstitute.dsm.util.NotificationUtil;
 
 public class PatchFactory {
